move sendMsg to `class` folder, comment cleanup 🧹
