+++ autoresponses/extractVids.js
@@ -4,10 +4,10 @@ const ytdlVids = require('../modules/ytdlVids')
 // Check for video links and send the raw video file using youtube-dl
 /** @param {import('discord.js').Message} message */
 module.exports = async message => {
-  // if it's a `;video` command then return immediately
+  // If it's a `;video` command then return immediately
   if (message.content.split(/ +/g)[0].toLowerCase() === ';video') return
 
-  // regex for getting links
+  // Regex for getting links
   const ytdlRegex = new RegExp('(?<=.|^|\\s)https:\\/\\/(www\\.)?' + // get "https://" and "www." if present
     '(' +
       'twitter\\.com\\/[A-z0-9_]{0,15}\\/status\\/[0-9]{5,}' + // get twitter.com/status/[user]/[id] links
@@ -22,16 +22,15 @@ module.exports = async message => {
     ')' +
     '(?=\\s|\\?|$)', 'g') // check if the next character is a space, a question mark or the end of the string
 
-  // get links from message and remove duplicates
+  // Get links from message and remove duplicates
   const links = [...new Set(message.content.match(ytdlRegex))]
-  // fetch each link
+  // Fetch each link
   links.forEach(link => ytdlVids(link, message.client).then(async files => {
-    // send video
+    // Send video
     if (!files || files.error) return
     while (true) {
       try {
-        await message.reply({ files, allowedMentions: { repliedUser: false } })
-        return
+        return await message.reply({ files, allowedMentions: { repliedUser: false } })
       } catch (err) {
         errorCatch(err, message.client)
       }
