+++ commands/image.js
@@ -209,7 +209,7 @@ const msgNotEnabled = `A moderator or admin hasn't enabled this command yet. \nT
 const msgDisableWarning = "When you disable the image command, you won't be able to use it until a moderator enables it back again. It will also remove ALL data regarding images sent so you wouldn't be able to search for them again. \nAre you really sure about this? Enter `" + prefix + 'image --disable server YES` to go ahead.'
 
 /**
- * @param {import('../modules/sendMsg')} msg
+ * @param {import('../class/sendMsg')} msg
  * @param {Array} args
  */
 exports.run = async (msg, args) => {
